Resolve problems
